Renamed limit orders updaters
