+++ src/cow-react/modules/limitOrders/containers/TradeButtons/index.tsx
@@ -15,7 +15,7 @@ import { PriceImpact } from 'hooks/usePriceImpact'
 import { useWrapCallback } from 'hooks/useWrapCallback'
 import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 import { transactionConfirmAtom } from '@cow/modules/swap/state/transactionConfirmAtom'
-import { ApplicationModal } from '@src/state/application/reducer'
+import { ApplicationModal } from 'state/application/reducer'
 import { useErrorModal } from 'hooks/useErrorMessageAndModal'
 import OperatorError from '@cow/api/gnosisProtocol/errors/OperatorError'
 import { CompatibilityIssuesWarning } from '@cow/modules/trade/pure/CompatibilityIssuesWarning'
