+++ src/cow-react/modules/limitOrders/hooks/useTradeFlowContext.ts
@@ -9,7 +9,7 @@ import { useAppData } from 'hooks/useAppData'
 import { LIMIT_ORDER_SLIPPAGE } from '@cow/modules/limitOrders/const/trade'
 import useENSAddress from 'hooks/useENSAddress'
 import { useLimitOrdersTradeState } from './useLimitOrdersTradeState'
-import { OrderClass } from '@src/custom/state/orders/actions'
+import { OrderClass } from 'state/orders/actions'
 import { useAtomValue } from 'jotai/utils'
 import { limitOrdersQuoteAtom } from '@cow/modules/limitOrders/state/limitOrdersQuoteAtom'
 import { useUpdateAtom } from 'jotai/utils'
