+++ src/cow-react/modules/limitOrders/index.ts
@@ -1,7 +1,7 @@
 export * from './containers/LimitOrdersWidget'
 export * from './containers/ChartWidget'
 export * from './containers/OrdersWidget'
-export * from './updaters/QuoteUpdater'
 export * from './updaters/MarketPriceUpdater'
+export * from './updaters/InitialPriceUpdater'
 export * from './pure/InfoPopup'
 export * from './state/limitOrdersAtom'
