+++ src/cow-react/modules/limitOrders/updaters/MarketPriceUpdater/hooks/useQuoteRequestParams.ts
@@ -7,7 +7,7 @@ import { useMemo } from 'react'
 import { useTypedValue } from '@cow/modules/limitOrders/hooks/useTypedValue'
 import { getAddress } from '@cow/utils/getAddress'
 import useENSAddress from 'hooks/useENSAddress'
-import { NATIVE_CURRENCY_BUY_ADDRESS } from '@src/custom/constants'
+import { NATIVE_CURRENCY_BUY_ADDRESS } from 'constants/index'
 
 export type LimitOrdersQuoteParams = Omit<FeeQuoteParams, 'validTo'> & { validTo?: number }
