--- src/cow-react/modules/limitOrders/updaters/QuoteUpdater/index.tsx
@@ -1,8 +1,0 @@
-import { useFetchMarketPrice } from './hooks/useFetchMarketPrice'
-
-// TODO: rename it to MarketPriceUpdater
-export function QuoteUpdater() {
-  useFetchMarketPrice()
-
-  return null
-}
